Implement forgot password and reset password functionality; add OTP generation and email sending for password recovery.
